Unified spacing style in server.js.
Changed the names of two endpoints:
    datasheetEager => datasheet, datasheet => update_datasheet

Hopeful fix for the double chart draw bug.  Instead of drawing the cache
then drawing a new one if it has been updated, we just always draw the
cache, which is now tracked.  We still make the request to the google
sheet to update the cache for the next load.

Removed d3 and crossfilter dependencies from package.json as they are
already in the bower.json